+++ changelog.md
@@ -3,13 +3,13 @@ This file contains all the notable changes done to the Ballerina Crypto package 
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+
 ## [Unreleased]
 
 ### Added
 
 - [Introduce new APIs to support password hashing and verification](https://github.com/ballerina-platform/ballerina-library/issues/2744)
-
-## [2.7.3] - 2025-01-16
 
 ### Added